@@ -22,15 +22,18 @@
     }
   },
   "require": {
-    "php": ">=7.3",
-    "nikic/php-parser": "^4.0"
+    "php": ">=8.2",
+    "nikic/php-parser": "^5.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "^7.0"
+    "phpunit/phpunit": "^11.0.7",
+    "tracy/tracy": "^2.10",
+    "rector/rector": "^1.0",
+    "rector/rector-php-parser": "^0.14.0"
   },
   "extra": {
     "branch-alias": {
-      "dev-master": "3.x-dev"
+      "dev-master": "4.x-dev"
     }
   }
 }
