@@ -12,9 +12,8 @@ declare(strict_types=1);
 
 namespace Go\ParserReflection\NodeVisitor;
 
-use Go\ParserReflection\ValueResolver\NodeExpressionResolver;
+use Go\ParserReflection\Resolver\NodeExpressionResolver;
 use PhpParser\Node;
-use PhpParser\NodeTraverser;
 use PhpParser\NodeVisitorAbstract;
 
 /**
@@ -24,19 +23,17 @@ class StaticVariablesCollector extends NodeVisitorAbstract
 {
     /**
      * Reflection context, eg. ReflectionClass, ReflectionMethod, etc
-     *
-     * @var mixed
      */
-    private $context;
+    private mixed $context;
 
-    private $staticVariables = [];
+    private array $staticVariables = [];
 
     /**
      * Default constructor
      *
      * @param mixed $context Reflection context, eg. ReflectionClass, ReflectionMethod, etc
      */
-    public function __construct($context)
+    public function __construct(mixed $context)
     {
         $this->context = $context;
     }
@@ -48,7 +45,7 @@ class StaticVariablesCollector extends NodeVisitorAbstract
     {
         // There may be internal closures, we do not need to look at them
         if ($node instanceof Node\Expr\Closure) {
-            return NodeTraverser::DONT_TRAVERSE_CHILDREN;
+            return self::DONT_TRAVERSE_CHILDREN;
         }
 
         if ($node instanceof Node\Stmt\Static_) {
