@@ -12,26 +12,30 @@ declare(strict_types=1);
 namespace Go\ParserReflection;
 
 use Closure;
+use Go\ParserReflection\Traits\AttributeResolverTrait;
 use Go\ParserReflection\Traits\InternalPropertiesEmulationTrait;
 use Go\ParserReflection\Traits\ReflectionFunctionLikeTrait;
+use JetBrains\PhpStorm\Deprecated;
 use PhpParser\Node\Stmt\Function_;
 use ReflectionFunction as BaseReflectionFunction;
 
 /**
  * AST-based reflection for function
+ * @see \Go\ParserReflection\ReflectionFunctionTest
  */
 class ReflectionFunction extends BaseReflectionFunction
 {
     use InternalPropertiesEmulationTrait;
     use ReflectionFunctionLikeTrait;
+    use AttributeResolverTrait;
 
     /**
      * Initializes reflection instance for given AST-node
      *
-     * @param string|Closure $functionName The name of the function to reflect or a closure.
+     * @param string $functionName The name of the function to reflect.
      * @param Function_ $functionNode Function node AST
      */
-    public function __construct($functionName, Function_ $functionNode)
+    public function __construct(string $functionName, Function_ $functionNode)
     {
         $namespaceParts = explode('\\', $functionName);
         // Remove the last one part with function name
@@ -67,7 +71,7 @@ class ReflectionFunction extends BaseReflectionFunction
     /**
      * {@inheritDoc}
      */
-    public function getClosure()
+    public function getClosure(): \Closure
     {
         $this->initializeInternalReflection();
 
@@ -77,17 +81,17 @@ class ReflectionFunction extends BaseReflectionFunction
     /**
      * {@inheritDoc}
      */
-    public function invoke($args = null)
+    public function invoke(mixed ...$args): mixed
     {
         $this->initializeInternalReflection();
 
-        return parent::invoke(...func_get_args());
+        return parent::invoke(...$args);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function invokeArgs(array $args)
+    public function invokeArgs(array $args): mixed
     {
         $this->initializeInternalReflection();
 
@@ -100,17 +104,16 @@ class ReflectionFunction extends BaseReflectionFunction
      * Only internal functions can be disabled using disable_functions directive.
      * User-defined functions are unaffected.
      */
-    public function isDisabled()
+    #[Deprecated('ReflectionFunction::isDisabled() is deprecated', since: "8.0")]
+    public function isDisabled(): bool
     {
         return false;
     }
 
     /**
      * Returns textual representation of function
-     *
-     * @return string
      */
-    public function __toString()
+    public function __toString(): string
     {
         $paramFormat      = ($this->getNumberOfParameters() > 0) ? "\n\n  - Parameters [%d] {%s\n  }" : '';
         $reflectionFormat = "%sFunction [ <user> function %s ] {\n  @@ %s %d - %d{$paramFormat}\n}\n";
@@ -123,9 +126,7 @@ class ReflectionFunction extends BaseReflectionFunction
             $this->getStartLine(),
             $this->getEndLine(),
             count($this->getParameters()),
-            array_reduce($this->getParameters(), static function ($str, ReflectionParameter $param) {
-                return $str . "\n    " . $param;
-            }, '')
+            array_reduce($this->getParameters(), static fn($str, ReflectionParameter $param) => $str . "\n    " . $param, '')
         );
     }
 
