@@ -11,79 +11,80 @@ declare(strict_types=1);
 
 namespace Go\ParserReflection;
 
+use Go\ParserReflection\Traits\AttributeResolverTrait;
 use Go\ParserReflection\Traits\InternalPropertiesEmulationTrait;
-use Go\ParserReflection\ValueResolver\NodeExpressionResolver;
+use Go\ParserReflection\Resolver\NodeExpressionResolver;
+use Go\ParserReflection\Resolver\TypeExpressionResolver;
+use JetBrains\PhpStorm\Deprecated;
+use PhpParser\Node\Expr;
+use PhpParser\Node\Expr\Array_;
+use PhpParser\Node\Expr\BinaryOp\Concat;
+use PhpParser\Node\Expr\ClassConstFetch;
+use PhpParser\Node\Expr\ConstFetch;
 use PhpParser\Node\Identifier;
 use PhpParser\Node\Name;
 use PhpParser\Node\NullableType;
 use PhpParser\Node\Param;
+use PhpParser\PrettyPrinter\Standard;
 use ReflectionFunctionAbstract;
 use ReflectionParameter as BaseReflectionParameter;
 
 /**
  * AST-based reflection for method/function parameter
+ * @see \Go\ParserReflection\ReflectionParameterTest
  */
 class ReflectionParameter extends BaseReflectionParameter
 {
     use InternalPropertiesEmulationTrait;
+    use AttributeResolverTrait;
 
     /**
      * Reflection function or method
-     *
-     * @var ReflectionFunctionAbstract
      */
-    private $declaringFunction;
+    private ReflectionFunctionAbstract $declaringFunction;
 
     /**
      * Stores the default value for node (if present)
-     *
-     * @var mixed
      */
-    private $defaultValue;
+    private mixed $defaultValue;
 
     /**
      * Whether or not default value is constant
-     *
-     * @var bool
      */
-    private $isDefaultValueConstant = false;
+    private bool $isDefaultValueConstant = false;
 
     /**
      * Name of the constant of default value
      *
-     * @var string
+     * @see $isDefaultValueConstant
      */
-    private $defaultValueConstantName;
+    private ?string $defaultValueConstantName;
 
     /**
      * Index of parameter in the list
-     *
-     * @var int
      */
-    private $parameterIndex;
+    private int $parameterIndex;
 
     /**
      * Concrete parameter node
-     *
-     * @var Param
      */
-    private $parameterNode;
+    private Param $parameterNode;
+
+    private bool $isDefaultValueConstExpr;
+
+    private ?string $defaultValueConstExpr;
+
+    private \ReflectionUnionType|\ReflectionNamedType|\ReflectionIntersectionType|null $type = null;
 
     /**
      * Initializes a reflection for the property
-     *
-     * @param string|array                $unusedFunctionName Name of the function/method
-     * @param string                      $parameterName      Name of the parameter to reflect
-     * @param ?Param                      $parameterNode      Parameter definition node
-     * @param int                         $parameterIndex     Index of parameter
-     * @param ?ReflectionFunctionAbstract $declaringFunction
      */
     public function __construct(
-        $unusedFunctionName,
-        $parameterName,
-        Param $parameterNode = null,
-        $parameterIndex = 0,
-        ReflectionFunctionAbstract $declaringFunction = null
+        string|array $unusedFunctionName,
+        string $parameterName,
+        Param $parameterNode,
+        int $parameterIndex,
+        ReflectionFunctionAbstract $declaringFunction
     ) {
         // Let's unset original read-only property to have a control over it via __get
         unset($this->name);
@@ -92,27 +93,38 @@ class ReflectionParameter extends BaseReflectionParameter
         $this->parameterIndex    = $parameterIndex;
         $this->declaringFunction = $declaringFunction;
 
-        if ($this->isDefaultValueAvailable()) {
-            if ($declaringFunction instanceof \ReflectionMethod) {
-                $context = $declaringFunction->getDeclaringClass();
-            } else {
-                $context = $declaringFunction;
-            }
+        if ($declaringFunction instanceof \ReflectionMethod) {
+            $context = $declaringFunction->getDeclaringClass();
+        } else {
+            $context = $declaringFunction;
+        }
 
+        if ($this->isDefaultValueAvailable()) {
             $expressionSolver = new NodeExpressionResolver($context);
             $expressionSolver->process($this->parameterNode->default);
+
             $this->defaultValue             = $expressionSolver->getValue();
             $this->isDefaultValueConstant   = $expressionSolver->isConstant();
             $this->defaultValueConstantName = $expressionSolver->getConstantName();
+            $this->isDefaultValueConstExpr  = $expressionSolver->isConstExpression();
+            $this->defaultValueConstExpr    = $expressionSolver->getConstExpression();
+        }
+
+        if ($this->hasType()) {
+            // If we have null value, this handled internally as nullable type too
+            $hasDefaultNull = $this->isDefaultValueAvailable() && $this->getDefaultValue() === null;
+
+            $typeResolver = new TypeExpressionResolver($this->getDeclaringClass());
+            $typeResolver->process($this->parameterNode->type, $hasDefaultNull);
+
+            $this->type = $typeResolver->getType();
         }
     }
 
     /**
      * Returns an AST-node for parameter
-     *
-     * @return Param
      */
-    public function getNode()
+    public function getNode(): Param
     {
         return $this->parameterNode;
     }
@@ -129,26 +141,22 @@ class ReflectionParameter extends BaseReflectionParameter
 
     /**
      * Returns string representation of this parameter.
-     *
-     * @return string
      */
-    public function __toString()
+    public function __toString(): string
     {
         $parameterType   = $this->getType();
         $isOptional      = $this->isOptional();
         $hasDefaultValue = $this->isDefaultValueAvailable();
         $defaultValue    = '';
         if ($hasDefaultValue) {
-            $defaultValue = $this->getDefaultValue();
-            if (is_string($defaultValue) && strlen($defaultValue) > 15) {
-                $defaultValue = substr($defaultValue, 0, 15) . '...';
+            // For constant fetch expressions, PHP renders now expression
+            if ($this->isDefaultValueConstExpr) {
+                $defaultValue = $this->defaultValueConstExpr;
+            } elseif ($this->isDefaultValueConstant){
+                $defaultValue = $this->defaultValueConstantName;
+            } else {
+                $defaultValue = var_export($this->getDefaultValue(), true);
             }
-            /* @see https://3v4l.org/DJOEb for behaviour changes */
-            if (is_float($defaultValue) && fmod($defaultValue, 1.0) === 0.0) {
-                $defaultValue = (int)$defaultValue;
-            }
-
-            $defaultValue = str_replace('\\\\', '\\', var_export($defaultValue, true));
         }
 
         return sprintf(
@@ -166,25 +174,20 @@ class ReflectionParameter extends BaseReflectionParameter
     /**
      * {@inheritDoc}
      */
-    public function allowsNull()
+    public function allowsNull(): bool
     {
-        // Enable 7.1 nullable types support
-        if ($this->parameterNode->type instanceof NullableType) {
+        // All non-typed parameters allows null by default
+        if (!$this->hasType()) {
             return true;
         }
 
-        $hasDefaultNull = $this->isDefaultValueAvailable() && $this->getDefaultValue() === null;
-        if ($hasDefaultNull) {
-            return true;
-        }
-
-        return !isset($this->parameterNode->type);
+        return $this->type->allowsNull();
     }
 
     /**
      * {@inheritDoc}
      */
-    public function canBePassedByValue()
+    public function canBePassedByValue(): bool
     {
         return !$this->isPassedByReference();
     }
@@ -192,10 +195,20 @@ class ReflectionParameter extends BaseReflectionParameter
     /**
      * @inheritDoc
      */
-    public function getClass()
+    #[Deprecated(reason: "Use ReflectionParameter::getType() and the ReflectionType APIs should be used instead.", since: "8.0")]
+    public function getClass(): ?\ReflectionClass
     {
         $parameterType = $this->parameterNode->type;
+        if ($parameterType instanceof Identifier && $parameterType->name === 'iterable') {
+            // This is how PHP represents iterable pseudo-class
+            $parameterType = new Name\FullyQualified(\Traversable::class);
+        }
         if ($parameterType instanceof Name) {
+            // If we have resolved type name, we should use it instead
+            if ($parameterType->hasAttribute('resolvedName')) {
+                $parameterType = $parameterType->getAttribute('resolvedName');
+            }
+
             if (!$parameterType instanceof Name\FullyQualified) {
                 $parameterTypeName = $parameterType->toString();
 
@@ -222,7 +235,7 @@ class ReflectionParameter extends BaseReflectionParameter
     /**
      * {@inheritDoc}
      */
-    public function getDeclaringClass()
+    public function getDeclaringClass(): ?\ReflectionClass
     {
         if ($this->declaringFunction instanceof \ReflectionMethod) {
             return $this->declaringFunction->getDeclaringClass();
@@ -234,7 +247,7 @@ class ReflectionParameter extends BaseReflectionParameter
     /**
      * {@inheritDoc}
      */
-    public function getDeclaringFunction()
+    public function getDeclaringFunction(): ReflectionFunctionAbstract
     {
         return $this->declaringFunction;
     }
@@ -242,7 +255,7 @@ class ReflectionParameter extends BaseReflectionParameter
     /**
      * {@inheritDoc}
      */
-    public function getDefaultValue()
+    public function getDefaultValue(): mixed
     {
         if (!$this->isDefaultValueAvailable()) {
             throw new ReflectionException('Internal error: Failed to retrieve the default value');
@@ -254,7 +267,7 @@ class ReflectionParameter extends BaseReflectionParameter
     /**
      * {@inheritDoc}
      */
-    public function getDefaultValueConstantName()
+    public function getDefaultValueConstantName(): null|string
     {
         if (!$this->isDefaultValueAvailable()) {
             throw new ReflectionException('Internal error: Failed to retrieve the default value');
@@ -266,7 +279,7 @@ class ReflectionParameter extends BaseReflectionParameter
     /**
      * @inheritDoc
      */
-    public function getName()
+    public function getName(): string
     {
         return (string)$this->parameterNode->var->name;
     }
@@ -274,7 +287,7 @@ class ReflectionParameter extends BaseReflectionParameter
     /**
      * {@inheritDoc}
      */
-    public function getPosition()
+    public function getPosition(): int
     {
         return $this->parameterIndex;
     }
@@ -282,43 +295,24 @@ class ReflectionParameter extends BaseReflectionParameter
     /**
      * @inheritDoc
      */
-    public function getType()
+    public function getType(): \ReflectionNamedType|\ReflectionUnionType|\ReflectionIntersectionType|null
     {
-        $isBuiltin     = false;
-        $parameterType = $this->parameterNode->type;
-        if ($parameterType instanceof NullableType) {
-            $parameterType = $parameterType->type;
-        }
-
-        $allowsNull = $this->allowsNull();
-        if ($parameterType instanceof Identifier) {
-            $isBuiltin = true;
-            $parameterType = $parameterType->toString();
-        } elseif (is_object($parameterType)) {
-            $parameterType = $parameterType->toString();
-        } elseif (is_string($parameterType)) {
-            $isBuiltin = true;
-        } else {
-            return null;
-        }
-
-        return new ReflectionNamedType($parameterType, $allowsNull, $isBuiltin);
+        return $this->type;
     }
 
     /**
      * @inheritDoc
      */
-    public function hasType()
+    public function hasType(): bool
     {
-        $hasType = isset($this->parameterNode->type);
-
-        return $hasType;
+        return isset($this->parameterNode->type);
     }
 
     /**
      * @inheritDoc
      */
-    public function isArray()
+    #[Deprecated(reason: "Use ReflectionParameter::getType() instead.", since: "8.0")]
+    public function isArray(): bool
     {
         $type = $this->parameterNode->type;
 
@@ -328,7 +322,8 @@ class ReflectionParameter extends BaseReflectionParameter
     /**
      * @inheritDoc
      */
-    public function isCallable()
+    #[Deprecated(reason: "Use ReflectionParameter::getType() instead.", since: "8.0")]
+    public function isCallable(): bool
     {
         $type = $this->parameterNode->type;
 
@@ -338,15 +333,15 @@ class ReflectionParameter extends BaseReflectionParameter
     /**
      * @inheritDoc
      */
-    public function isDefaultValueAvailable()
+    public function isDefaultValueAvailable(): bool
     {
-        return isset($this->parameterNode->default);
+        return isset($this->parameterNode->default) && $this->allSiblingsAreOptional();
     }
 
     /**
      * {@inheritDoc}
      */
-    public function isDefaultValueConstant()
+    public function isDefaultValueConstant(): bool
     {
         return $this->isDefaultValueConstant;
     }
@@ -354,44 +349,45 @@ class ReflectionParameter extends BaseReflectionParameter
     /**
      * {@inheritDoc}
      */
-    public function isOptional()
+    public function isOptional(): bool
     {
-        return $this->isVariadic() || ($this->isDefaultValueAvailable() && $this->haveSiblingsDefaultValues());
+        return $this->isVariadic() || $this->isDefaultValueAvailable();
     }
 
     /**
      * @inheritDoc
      */
-    public function isPassedByReference()
+    public function isPassedByReference(): bool
     {
-        return (bool) $this->parameterNode->byRef;
+        return $this->parameterNode->byRef;
     }
 
     /**
      * @inheritDoc
      */
-    public function isVariadic()
+    public function isPromoted(): bool
     {
-        return (bool) $this->parameterNode->variadic;
+        return $this->parameterNode->isPromoted();
     }
 
     /**
-     * Returns if all following parameters have a default value definition.
-     *
-     * @return bool
-     * @throws ReflectionException If could not fetch declaring function reflection
+     * @inheritDoc
      */
-    protected function haveSiblingsDefaultValues()
+    public function isVariadic(): bool
     {
-        $function = $this->getDeclaringFunction();
-        if (null === $function) {
-            throw new ReflectionException('Could not get the declaring function reflection.');
-        }
+        return $this->parameterNode->variadic;
+    }
 
-        /** @var BaseReflectionParameter[] $remainingParameters */
-        $remainingParameters = array_slice($function->getParameters(), $this->parameterIndex + 1);
-        foreach ($remainingParameters as $reflectionParameter) {
-            if (!$reflectionParameter->isDefaultValueAvailable()) {
+    /**
+     * Returns true if all following parameters are optional (either have values or variadic)
+     */
+    private function allSiblingsAreOptional(): bool
+    {
+        // start from PHP 8.1, isDefaultValueAvailable() returns false if next parameter is required
+        // see https://github.com/php/php-src/issues/8090
+        $parameters = $this->declaringFunction->getNode()->getParams();
+        for ($nextParamIndex = $this->parameterIndex + 1; $nextParamIndex < count($parameters); ++$nextParamIndex) {
+            if (!isset($parameters[$nextParamIndex]->default) && !$parameters[$nextParamIndex]->variadic) {
                 return false;
             }
         }
