@@ -15,8 +15,9 @@ namespace Go\ParserReflection\Traits;
 
 use Go\ParserReflection\NodeVisitor\GeneratorDetector;
 use Go\ParserReflection\NodeVisitor\StaticVariablesCollector;
-use Go\ParserReflection\ReflectionNamedType;
+use Go\ParserReflection\ReflectionMethod;
 use Go\ParserReflection\ReflectionParameter;
+use Go\ParserReflection\Resolver\TypeExpressionResolver;
 use PhpParser\Node\Expr\Closure;
 use PhpParser\Node\FunctionLike;
 use PhpParser\Node\Identifier;
@@ -24,6 +25,8 @@ use PhpParser\Node\NullableType;
 use PhpParser\Node\Stmt\ClassMethod;
 use PhpParser\Node\Stmt\Function_;
 use PhpParser\NodeTraverser;
+use ReflectionException;
+use ReflectionExtension;
 
 /**
  * General trait for all function-like reflections
@@ -32,27 +35,22 @@ trait ReflectionFunctionLikeTrait
 {
     use InitializationTrait;
 
-    /**
-     * @var FunctionLike
-     */
-    protected $functionLikeNode;
+    protected FunctionLike|Function_|ClassMethod $functionLikeNode;
 
     /**
      * Namespace name
-     *
-     * @var string
      */
-    protected $namespaceName = '';
+    protected string $namespaceName = '';
 
     /**
-     * @var array|ReflectionParameter[]
+     * @var ReflectionParameter[]
      */
-    protected $parameters;
+    protected array $parameters;
 
     /**
      * {@inheritDoc}
      */
-    public function getClosureScopeClass()
+    public function getClosureScopeClass(): ?\ReflectionClass
     {
         $this->initializeInternalReflection();
 
@@ -62,49 +60,43 @@ trait ReflectionFunctionLikeTrait
     /**
      * {@inheritDoc}
      */
-    public function getClosureThis()
+    public function getClosureThis(): ?object
     {
         $this->initializeInternalReflection();
 
         return parent::getClosureThis();
     }
 
-    public function getDocComment()
+    public function getDocComment(): string|false
     {
         $docComment = $this->functionLikeNode->getDocComment();
 
         return $docComment ? $docComment->getText() : false;
     }
 
-    public function getEndLine()
+    public function getEndLine(): int|false
     {
-        return $this->functionLikeNode->getAttribute('endLine');
+        if ($this->functionLikeNode->hasAttribute('endLine')) {
+            return $this->functionLikeNode->getAttribute('endLine');
+        }
+
+        return false;
     }
 
-    public function getExtension()
+    public function getExtension(): ?ReflectionExtension
     {
         return null;
     }
 
-    public function getExtensionName()
+    public function getExtensionName(): string|false
     {
         return false;
     }
 
-    public function getFileName()
+    public function getFileName(): string|false
     {
-        return $this->functionLikeNode->getAttribute('fileName');
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getName()
-    {
-        if ($this->functionLikeNode instanceof Function_ || $this->functionLikeNode instanceof ClassMethod) {
-            $functionName = $this->functionLikeNode->name->toString();
-
-            return $this->namespaceName ? $this->namespaceName . '\\' . $functionName : $functionName;
+        if ($this->functionLikeNode->hasAttribute('fileName')) {
+            return $this->functionLikeNode->getAttribute('fileName');
         }
 
         return false;
@@ -113,7 +105,21 @@ trait ReflectionFunctionLikeTrait
     /**
      * {@inheritDoc}
      */
-    public function getNamespaceName()
+    public function getName(): string
+    {
+        if ($this->functionLikeNode instanceof Function_ || $this->functionLikeNode instanceof ClassMethod) {
+            $functionName = $this->functionLikeNode->name->toString();
+
+            return $this->namespaceName ? $this->namespaceName . '\\' . $functionName : $functionName;
+        }
+
+        return '';
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getNamespaceName(): string
     {
         return $this->namespaceName;
     }
@@ -122,10 +128,8 @@ trait ReflectionFunctionLikeTrait
      * Get the number of parameters that a function defines, both optional and required.
      *
      * @link http://php.net/manual/en/reflectionfunctionabstract.getnumberofparameters.php
-     *
-     * @return int
      */
-    public function getNumberOfParameters()
+    public function getNumberOfParameters(): int
     {
         return count($this->functionLikeNode->getParams());
     }
@@ -134,10 +138,8 @@ trait ReflectionFunctionLikeTrait
      * Get the number of required parameters that a function defines.
      *
      * @link http://php.net/manual/en/reflectionfunctionabstract.getnumberofrequiredparameters.php
-     *
-     * @return int
      */
-    public function getNumberOfRequiredParameters()
+    public function getNumberOfRequiredParameters(): int
     {
         $requiredParameters = 0;
         foreach ($this->getParameters() as $parameter) {
@@ -152,7 +154,7 @@ trait ReflectionFunctionLikeTrait
     /**
      * {@inheritDoc}
      */
-    public function getParameters()
+    public function getParameters(): array
     {
         if (!isset($this->parameters)) {
             $parameters = [];
@@ -178,54 +180,52 @@ trait ReflectionFunctionLikeTrait
     /**
      * Gets the specified return type of a function
      *
-     * @return \ReflectionType
-     *
      * @link http://php.net/manual/en/reflectionfunctionabstract.getreturntype.php
      */
-    public function getReturnType()
+    public function getReturnType(): \ReflectionNamedType|\ReflectionUnionType|\ReflectionIntersectionType|null
     {
-        $isBuiltin  = false;
-        $returnType = $this->functionLikeNode->getReturnType();
-        $isNullable = $returnType instanceof NullableType;
+        if ($this->hasReturnType()) {
+            $typeResolver = new TypeExpressionResolver($this);
+            $typeResolver->process($this->functionLikeNode->getReturnType(), false);
 
-        if ($isNullable) {
-            $returnType = $returnType->type;
-        }
-        if ($returnType instanceof Identifier) {
-            $isBuiltin  = true;
-            $returnType = $returnType->toString();
-        } elseif (is_object($returnType)) {
-            $returnType = $returnType->toString();
-        } elseif (is_string($returnType)) {
-            $isBuiltin = true;
-        } else {
-            return null;
+            return $typeResolver->getType();
         }
 
-        return new ReflectionNamedType($returnType, $isNullable, $isBuiltin);
+        return null;
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getShortName()
+    public function getShortName(): string
     {
         if ($this->functionLikeNode instanceof Function_ || $this->functionLikeNode instanceof ClassMethod) {
             return $this->functionLikeNode->name->toString();
         }
 
-        return false;
+        throw new ReflectionException('unable to get short name');
     }
 
-    public function getStartLine()
+    public function getStartLine(): int|false
     {
-        return $this->functionLikeNode->getAttribute('startLine');
+        if ($this->functionLikeNode->attrGroups !== []) {
+            $attrGroups = $this->functionLikeNode->attrGroups;
+            $lastAttrGroupsEndLine = end($attrGroups)->getAttribute('endLine');
+
+            return $lastAttrGroupsEndLine + 1;
+        }
+
+        if ($this->functionLikeNode->hasAttribute('startLine')) {
+            return $this->functionLikeNode->getAttribute('startLine');
+        }
+
+        return false;
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getStaticVariables()
+    public function getStaticVariables(): array
     {
         $nodeTraverser      = new NodeTraverser();
         $variablesCollector = new StaticVariablesCollector($this);
@@ -240,11 +240,10 @@ trait ReflectionFunctionLikeTrait
     /**
      * Checks if the function has a specified return type
      *
-     * @return bool
      *
      * @link http://php.net/manual/en/reflectionfunctionabstract.hasreturntype.php
      */
-    public function hasReturnType()
+    public function hasReturnType(): bool
     {
         $returnType = $this->functionLikeNode->getReturnType();
 
@@ -254,7 +253,7 @@ trait ReflectionFunctionLikeTrait
     /**
      * {@inheritDoc}
      */
-    public function inNamespace()
+    public function inNamespace(): bool
     {
         return !empty($this->namespaceName);
     }
@@ -262,7 +261,7 @@ trait ReflectionFunctionLikeTrait
     /**
      * {@inheritDoc}
      */
-    public function isClosure()
+    public function isClosure(): bool
     {
         return $this->functionLikeNode instanceof Closure;
     }
@@ -270,7 +269,7 @@ trait ReflectionFunctionLikeTrait
     /**
      * {@inheritDoc}
      */
-    public function isDeprecated()
+    public function isDeprecated(): bool
     {
         // user-land method/function/closure can not be deprecated
         return false;
@@ -279,7 +278,7 @@ trait ReflectionFunctionLikeTrait
     /**
      * {@inheritDoc}
      */
-    public function isGenerator()
+    public function isGenerator(): bool
     {
         $nodeTraverser = new NodeTraverser();
         $nodeDetector  = new GeneratorDetector();
@@ -294,25 +293,27 @@ trait ReflectionFunctionLikeTrait
     /**
      * {@inheritDoc}
      */
-    public function isInternal()
+    public function isInternal(): bool
     {
-        // never can be an internal method
-        return false;
+        // never can be an internal method, except for the Enum magic methods
+        $isEnumMethod = $this instanceof ReflectionMethod && $this->getDeclaringClass()->isEnum();
+        return $isEnumMethod && in_array($this->getName(), ['cases', 'tryFrom', 'from']);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function isUserDefined()
+    public function isUserDefined(): bool
     {
-        // always defined by user, because we parse the source code
-        return true;
+        // always user-defined method, except for the Enum magic methods
+        $isEnumMethod = $this instanceof ReflectionMethod && $this->getDeclaringClass()->isEnum();
+        return !($isEnumMethod && in_array($this->getName(), ['cases', 'tryFrom', 'from']));
     }
 
     /**
      * {@inheritDoc}
      */
-    public function isVariadic()
+    public function isVariadic(): bool
     {
         foreach ($this->getParameters() as $parameter) {
             if ($parameter->isVariadic()) {
@@ -326,7 +327,7 @@ trait ReflectionFunctionLikeTrait
     /**
      * {@inheritDoc}
      */
-    public function returnsReference()
+    public function returnsReference(): bool
     {
         return $this->functionLikeNode->returnsByRef();
     }
