@@ -9,12 +9,9 @@ class ReflectionTypeTest extends TestCase
 {
     protected function setUp(): void
     {
-        if (PHP_VERSION_ID >= 70000) {
-            include_once (__DIR__ . '/Stub/FileWithClasses70.php');
-        }
-        if (PHP_VERSION_ID >= 70100) {
-            include_once (__DIR__ . '/Stub/FileWithClasses71.php');
-        }
+        include_once (__DIR__ . '/Stub/FileWithClasses70.php');
+        include_once (__DIR__ . '/Stub/FileWithClasses71.php');
+        include_once (__DIR__ . '/Stub/FileWithClasses80.php');
     }
 
     /**
@@ -23,24 +20,19 @@ class ReflectionTypeTest extends TestCase
      * We're already testing it with Go\ParserReflection\ReflectionType
      * elsewhere.
      */
-    public function testTypeConvertToDisplayTypeWithNativeType()
+    public function testTypeConvertToDisplayTypeWithNativeType(): void
     {
-        $nativeClassRef = new \ReflectionClass('Go\\ParserReflection\\Stub\\ClassWithScalarTypeHints');
+        $nativeClassRef = new \ReflectionClass('Go\\ParserReflection\\Stub\\ClassWithPhp70ScalarTypeHints');
         $nativeMethodRef = $nativeClassRef->getMethod('acceptsDefaultString');
-        $this->assertEquals(\ReflectionMethod::class, get_class($nativeMethodRef));
+        $this->assertInstanceOf(\ReflectionMethod::class, $nativeMethodRef);
         $nativeParamRefArr = $nativeMethodRef->getParameters();
         $this->assertCount(2, $nativeParamRefArr);
-        $this->assertEquals(\ReflectionParameter::class, get_class($nativeParamRefArr[0]));
+        $this->assertInstanceOf(\ReflectionParameter::class, $nativeParamRefArr[0]);
         $nativeTypeRef = $nativeParamRefArr[0]->getType();
-        // TODO: To prevent deprecation error in tests
-        if (PHP_VERSION_ID < 70400) {
-            $this->assertEquals('string', (string)$nativeTypeRef);
-        } else {
-            $this->assertEquals('string', $nativeTypeRef->getName());
-        }
-        $this->assertNotContains('\\', get_class($nativeTypeRef));
+        $this->assertSame('string', $nativeTypeRef->getName());
+        $this->assertStringNotContainsString('\\', get_class($nativeTypeRef));
         $this->assertInstanceOf(\ReflectionType::class, $nativeTypeRef);
-        $this->assertEquals('string', \Go\ParserReflection\ReflectionType::convertToDisplayType($nativeTypeRef));
+        $this->assertSame('string', \Go\ParserReflection\ReflectionType::convertToDisplayType($nativeTypeRef));
     }
 
     /**
@@ -49,24 +41,19 @@ class ReflectionTypeTest extends TestCase
      * We're already testing it with Go\ParserReflection\ReflectionType
      * elsewhere.
      */
-    public function testTypeConvertToDisplayTypeWithNullableNativeType()
+    public function testTypeConvertToDisplayTypeWithNullableNativeType(): void
     {
-        $nativeClassRef = new \ReflectionClass('Go\\ParserReflection\\Stub\\ClassWithNullableScalarTypeHints');
+        $nativeClassRef = new \ReflectionClass('Go\\ParserReflection\\Stub\\ClassWithPhp71NullableScalarTypeHints');
         $nativeMethodRef = $nativeClassRef->getMethod('acceptsDefaultString');
-        $this->assertEquals(\ReflectionMethod::class, get_class($nativeMethodRef));
+        $this->assertInstanceOf(\ReflectionMethod::class, $nativeMethodRef);
         $nativeParamRefArr = $nativeMethodRef->getParameters();
         $this->assertCount(2, $nativeParamRefArr);
-        $this->assertEquals(\ReflectionParameter::class, get_class($nativeParamRefArr[0]));
+        $this->assertInstanceOf(\ReflectionParameter::class, $nativeParamRefArr[0]);
         $nativeTypeRef = $nativeParamRefArr[0]->getType();
-        // TODO: To prevent deprecation error in tests
-        if (PHP_VERSION_ID < 70400) {
-            $this->assertEquals('string', (string)$nativeTypeRef);
-        } else {
-            $this->assertEquals('string', $nativeTypeRef->getName());
-        }
-        $this->assertNotContains('\\', get_class($nativeTypeRef));
+        $this->assertSame('string', $nativeTypeRef->getName());
+        $this->assertStringNotContainsString('\\', get_class($nativeTypeRef));
         $this->assertInstanceOf(\ReflectionType::class, $nativeTypeRef);
-        $this->assertEquals('string or NULL', \Go\ParserReflection\ReflectionType::convertToDisplayType($nativeTypeRef));
+        $this->assertSame('?string', \Go\ParserReflection\ReflectionType::convertToDisplayType($nativeTypeRef));
     }
 
     /**
@@ -75,24 +62,41 @@ class ReflectionTypeTest extends TestCase
      * We're already testing it with Go\ParserReflection\ReflectionType
      * elsewhere.
      */
-    public function testTypeConvertToDisplayTypeImplicitlyNullable()
+    public function testTypeConvertToDisplayTypeImplicitlyNullable(): void
     {
-        $nativeClassRef = new \ReflectionClass('Go\\ParserReflection\\Stub\\ClassWithScalarTypeHints');
+        $nativeClassRef = new \ReflectionClass('Go\\ParserReflection\\Stub\\ClassWithPhp70ScalarTypeHints');
         $nativeMethodRef = $nativeClassRef->getMethod('acceptsStringDefaultToNull');
-        $this->assertEquals(\ReflectionMethod::class, get_class($nativeMethodRef));
+        $this->assertInstanceOf(\ReflectionMethod::class, $nativeMethodRef);
         $nativeParamRefArr = $nativeMethodRef->getParameters();
         $this->assertCount(1, $nativeParamRefArr);
-        $this->assertEquals(\ReflectionParameter::class, get_class($nativeParamRefArr[0]));
+        $this->assertInstanceOf(\ReflectionParameter::class, $nativeParamRefArr[0]);
         $nativeTypeRef = $nativeParamRefArr[0]->getType();
         $this->assertTrue($nativeTypeRef->allowsNull());
-        // TODO: To prevent deprecation error in tests
-        if (PHP_VERSION_ID < 70400) {
-            $this->assertEquals('string', (string)$nativeTypeRef);
-        } else {
-            $this->assertEquals('string', $nativeTypeRef->getName());
-        }
-        $this->assertNotContains('\\', get_class($nativeTypeRef));
+        $this->assertSame('string', $nativeTypeRef->getName());
+        $this->assertStringNotContainsString('\\', get_class($nativeTypeRef));
         $this->assertInstanceOf(\ReflectionType::class, $nativeTypeRef);
-        $this->assertEquals('string or NULL', \Go\ParserReflection\ReflectionType::convertToDisplayType($nativeTypeRef));
+        $this->assertSame('?string', \Go\ParserReflection\ReflectionType::convertToDisplayType($nativeTypeRef));
+    }
+
+    /**
+     * Testing convertToDisplayType() with native \ReflectionType
+     *
+     * We're already testing it with Go\ParserReflection\ReflectionType
+     * elsewhere.
+     */
+    public function testTypeConvertToDisplayTypeImplicitlyUnionNullable(): void
+    {
+        $nativeClassRef = new \ReflectionClass('Go\\ParserReflection\\Stub\\ClassWithPhp80Features');
+        $nativeMethodRef = $nativeClassRef->getMethod('acceptsStringArrayDefaultToNull');
+        $this->assertInstanceOf(\ReflectionMethod::class, $nativeMethodRef);
+        $nativeParamRefArr = $nativeMethodRef->getParameters();
+        $this->assertCount(1, $nativeParamRefArr);
+        $this->assertInstanceOf(\ReflectionParameter::class, $nativeParamRefArr[0]);
+        $nativeTypeRef = $nativeParamRefArr[0]->getType();
+        $this->assertTrue($nativeTypeRef->allowsNull());
+        $this->assertSame('array|string|null', (string) $nativeTypeRef);
+        $this->assertStringNotContainsString('\\', get_class($nativeTypeRef));
+        $this->assertInstanceOf(\ReflectionType::class, $nativeTypeRef);
+        $this->assertSame('array|string|null', \Go\ParserReflection\ReflectionType::convertToDisplayType($nativeTypeRef));
     }
 }
