@@ -21,8 +21,6 @@ interface LocatorInterface
      * Returns a path to the file for given class name
      *
      * @param string $className Name of the class (with or without leading '\' FQCN)
-     *
-     * @return string|false Path to the file with given class or false if not found
      */
-    public function locateClass(string $className);
+    public function locateClass(string $className): false|string;
 }
