@@ -12,29 +12,34 @@ declare(strict_types=1);
 
 namespace Go\ParserReflection;
 
+use Go\ParserReflection\Traits\AttributeResolverTrait;
 use Go\ParserReflection\Traits\InternalPropertiesEmulationTrait;
 use Go\ParserReflection\Traits\ReflectionClassLikeTrait;
+use PhpParser\Node\Name;
 use PhpParser\Node\Name\FullyQualified;
 use PhpParser\Node\Stmt\ClassLike;
+use PhpParser\Node\Stmt\Enum_;
 use PhpParser\Node\Stmt\Interface_;
 use PhpParser\Node\Stmt\TraitUse;
 use ReflectionClass as InternalReflectionClass;
 
 /**
  * AST-based reflection class
+ * @see \Go\ParserReflection\ReflectionClassTest
  */
 class ReflectionClass extends InternalReflectionClass
 {
     use InternalPropertiesEmulationTrait;
     use ReflectionClassLikeTrait;
+    use AttributeResolverTrait;
 
     /**
      * Initializes reflection instance
      *
-     * @param string|object $argument      Class name or instance of object
+     * @param object|string $argument      Class name or instance of object
      * @param ?ClassLike    $classLikeNode AST node for class
      */
-    public function __construct($argument, ClassLike $classLikeNode = null)
+    public function __construct(object|string $argument, ?ClassLike $classLikeNode = null)
     {
         $fullClassName   = is_object($argument) ? get_class($argument) : ltrim($argument, '\\');
         $namespaceParts  = explode('\\', $fullClassName);
@@ -58,18 +63,31 @@ class ReflectionClass extends InternalReflectionClass
 
         $isInterface    = $classLikeNode instanceof Interface_;
         $interfaceField = $isInterface ? 'extends' : 'implements';
-        $hasInterfaces  = in_array($interfaceField, $classLikeNode->getSubNodeNames(), true);
-        $implementsList = $hasInterfaces ? $classLikeNode->$interfaceField : [];
-        if ($implementsList) {
+
+        $hasExplicitInterfaces = in_array($interfaceField, $classLikeNode->getSubNodeNames(), true);
+        $implementsList        = $hasExplicitInterfaces ? $classLikeNode->$interfaceField : [];
+
+        if (count($implementsList) > 0) {
             foreach ($implementsList as $implementNode) {
-                if ($implementNode instanceof FullyQualified) {
-                    $implementName = $implementNode->toString();
+                if ($implementNode instanceof Name && $implementNode->getAttribute('resolvedName') instanceof FullyQualified) {
+                    $implementName = $implementNode->getAttribute('resolvedName')->toString();
                     $interface     = interface_exists($implementName, false)
                         ? new parent($implementName)
                         : new static($implementName);
 
                     $interfaces[$implementName] = $interface;
                 }
+            }
+        }
+
+        // All Enum classes has implicit interface(s) added by PHP
+        if ($classLikeNode instanceof Enum_) {
+            // @see https://php.watch/versions/8.1/enums#enum-BackedEnum
+            $interfacesToAdd = isset($classLikeNode->scalarType)
+                ? [\UnitEnum::class, \BackedEnum::class] // PHP Uses exactly this order, not reversed by parent!
+                : [\UnitEnum::class];
+            foreach ($interfacesToAdd as $interfaceToAdd) {
+                $interfaces[$interfaceToAdd] = new parent($interfaceToAdd);
             }
         }
 
@@ -91,8 +109,8 @@ class ReflectionClass extends InternalReflectionClass
             foreach ($classLikeNode->stmts as $classLevelNode) {
                 if ($classLevelNode instanceof TraitUse) {
                     foreach ($classLevelNode->traits as $classTraitName) {
-                        if ($classTraitName instanceof FullyQualified) {
-                            $traitName          = $classTraitName->toString();
+                        if ($classTraitName instanceof Name && $classTraitName->getAttribute('resolvedName') instanceof FullyQualified) {
+                            $traitName          = $classTraitName->getAttribute('resolvedName')->toString();
                             $trait              = trait_exists($traitName, false)
                                 ? new parent($traitName)
                                 : new static($traitName);
@@ -127,8 +145,6 @@ class ReflectionClass extends InternalReflectionClass
 
     /**
      * Implementation of internal reflection initialization
-     *
-     * @return void
      */
     protected function __initialize(): void
     {
@@ -138,12 +154,11 @@ class ReflectionClass extends InternalReflectionClass
     /**
      * Create a ReflectionClass for a given class name.
      *
-     * @param string $className
-     *     The name of the class to create a reflection for.
+     * @param string $className The name of the class to create a reflection for.
      *
      * @return InternalReflectionClass The appropriate reflection object.
      */
-    protected function createReflectionForClass(string $className)
+    protected function createReflectionForClass(string $className): InternalReflectionClass
     {
         return class_exists($className, false) ? new parent($className) : new static($className);
     }
