@@ -5,10 +5,11 @@ namespace Go\ParserReflection\Locator;
 
 use PHPUnit\Framework\TestCase;
 use Go\ParserReflection\ReflectionClass;
+use Go\ParserReflection\ReflectionEngine;
 
 class ComposerLocatorTest extends TestCase
 {
-    public function testLocateClass()
+    public function testLocateClass(): void
     {
         $locator         = new ComposerLocator();
         $reflectionClass = new \ReflectionClass(ReflectionClass::class);
@@ -20,5 +21,13 @@ class ComposerLocatorTest extends TestCase
             $reflectionClass->getFileName(),
             $locator->locateClass('\\' . ReflectionClass::class)
         );
+    }
+
+    public function testLocateClassWithAttributes(): void
+    {
+        ReflectionEngine::init(new ComposerLocator());
+
+        $parsedClass = new \Go\ParserReflection\ReflectionClass(\Go\ParserReflection\Stub\RandomClassWithAttribute::class);
+        $this->assertIsArray($parsedClass->getAttributes());
     }
 }
