@@ -3,114 +3,92 @@ declare(strict_types=1);
 
 namespace Go\ParserReflection;
 
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
-use Go\ParserReflection\Stub\Foo;
-use Go\ParserReflection\Stub\SubFoo;
-use TestParametersForRootNsClass;
-use Go\ParserReflection\Stub\ClassWithPhp71Features;
 
-class ReflectionClassConstantTest extends TestCase
+class ReflectionClassConstantTest extends AbstractTestCase
 {
-    /**
-     * @var ReflectionFile
-     */
-    protected $parsedRefFile;
-
-    protected function setUp(): void
-    {
-        $this->setUpFile(__DIR__ . '/Stub/FileWithClasses71.php');
-    }
-
-    public function testGeneralInfoGetters()
-    {
-        $allNameGetters = [
-            'getDocComment',
-            'getModifiers',
-            'getName',
-            'getValue',
-            'isPrivate',
-            'isProtected',
-            'isPublic',
-            '__toString'
-        ];
-
-        foreach ($this->parsedRefFile->getFileNamespaces() as $fileNamespace) {
-            foreach ($fileNamespace->getClasses() as $refClass) {
-                $className = $refClass->getName();
-                foreach ($refClass->getReflectionConstants() as $refReflectionConstant) {
-                    $classConstantName = $refReflectionConstant->getName();
-                    $originalRefParameter = new \ReflectionClassConstant($className, $classConstantName);
-                    foreach ($allNameGetters as $getterName) {
-                        $expectedValue = $originalRefParameter->$getterName();
-                        $actualValue = $refReflectionConstant->$getterName();
-                        $this->assertSame(
-                            $expectedValue,
-                            $actualValue,
-                            "{$getterName}() for parameter {$className}::{$classConstantName} should be equal"
-                        );
-                    }
-                }
-            }
-        }
-    }
-
-    public function testGetClassConstantProperties()
-    {
-        $parsedNamespace = $this->parsedRefFile->getFileNamespace('Go\ParserReflection\Stub');
-        $parsedClass = $parsedNamespace->getClass(ClassWithPhp71Features::class);
-
-        $constant = $parsedClass->getReflectionConstant('PUBLIC_CONST_A');
-        $this->assertSame('PUBLIC_CONST_A', $constant->name);
-        $this->assertSame(ClassWithPhp71Features::class, $constant->class);
-    }
-
-    public function testGetClassConstant()
-    {
-        $parsedNamespace = $this->parsedRefFile->getFileNamespace('Go\ParserReflection\Stub');
-        $parsedClass = $parsedNamespace->getClass(ClassWithPhp71Features::class);
-
-        $classConstants = $parsedClass->getReflectionConstants();
-        $this->assertSame($classConstants[0], $parsedClass->getReflectionConstant('PUBLIC_CONST_A'));
-        $this->assertSame($classConstants[1], $parsedClass->getReflectionConstant('PUBLIC_CONST_B'));
-        $this->assertSame($classConstants[2], $parsedClass->getReflectionConstant('PROTECTED_CONST'));
-        $this->assertSame($classConstants[3], $parsedClass->getReflectionConstant('PRIVATE_CONST'));
-        $this->assertSame($classConstants[4], $parsedClass->getReflectionConstant('CALCULATED_CONST'));
-    }
-
-    public function testCoverAllMethods()
-    {
-        $allInternalMethods = get_class_methods(\ReflectionClassConstant::class);
-        $allMissedMethods = [];
-
-        foreach ($allInternalMethods as $internalMethodName) {
-            if ('export' === $internalMethodName) {
-                continue;
-            }
-            $refMethod = new \ReflectionMethod(ReflectionClassConstant::class, $internalMethodName);
-            $definerClass = $refMethod->getDeclaringClass()->getName();
-            if (strpos($definerClass, 'Go\\ParserReflection') !== 0) {
-                $allMissedMethods[] = $internalMethodName;
-            }
-        }
-
-        if ($allMissedMethods) {
-            $this->markTestIncomplete('Methods ' . implode(', ', $allMissedMethods) . ' are not implemented');
-        }
-    }
+    protected static string $reflectionClassToTest = \ReflectionClassConstant::class;
 
     /**
-     * Setups file for parsing
+     * Performs method-by-method comparison with original reflection
      *
-     * @param string $fileName File name to use
+     * @param ReflectionClass          $parsedClass Parsed class
+     * @param \ReflectionClassConstant $refClassConstant Method to analyze
+     * @param string                   $getterName Name of the reflection method to test
      */
-    private function setUpFile($fileName)
+    #[DataProvider('reflectionGetterDataProvider')]
+    public function testReflectionGetterParity(
+        ReflectionClass          $parsedClass,
+        \ReflectionClassConstant $refClassConstant,
+        string                   $getterName
+    ): void {
+        $constantName   = $refClassConstant->getName();
+        $className      = $parsedClass->getName();
+        $parsedConstant = $parsedClass->getReflectionConstant($constantName);
+
+        $expectedValue = $refClassConstant->$getterName();
+        $actualValue   = $parsedConstant->$getterName();
+        // I would like to completely stop maintaining the __toString method
+        if ($expectedValue !== $actualValue && $getterName === '__toString') {
+            $this->markTestSkipped("__toString for constant {$className}::{$constantName} is not equal:\n{$expectedValue}{$actualValue}");
+        }
+        if ($getterName === 'getValue' && $parsedClass->isEnum()) {
+            $this->markTestSkipped("getValue() for Enum cases could not be resolved, see https://github.com/goaop/parser-reflection/issues/132");
+        }
+        $this->assertSame(
+            $expectedValue,
+            $actualValue,
+            "$getterName() for constant $className::$constantName() should be equal"
+        );
+    }
+
+    /**
+     * Provides full test-case list in the form [ReflectionClass, \ReflectionClassConstant, getter name to check]
+     */
+    public static function reflectionGetterDataProvider(): \Generator
     {
-        $fileName = stream_resolve_include_path($fileName);
-        $fileNode = ReflectionEngine::parseFile($fileName);
+        $allNameGetters = self::getGettersToCheck();
+        foreach (self::classConstantsDataProvider() as $prefix => [$parsedClass, $classReflectionConstant]) {
+            foreach ($allNameGetters as $getterName) {
+                yield $prefix . ', ' . $getterName => [
+                    $parsedClass,
+                    $classReflectionConstant,
+                    $getterName
+                ];
+            }
+        }
+    }
 
-        $reflectionFile = new ReflectionFile($fileName, $fileNode);
-        $this->parsedRefFile = $reflectionFile;
+    /**
+     * Provides full test-case list in the form [ParsedClass, \ReflectionClassConstant to check]
+     */
+    public static function classConstantsDataProvider(): \Generator
+    {
+        foreach (self::classesDataProvider() as $prefix => [$parsedClass, $refClass]) {
+            foreach ($refClass->getReflectionConstants() as $reflectionConstant) {
+                $fullConstantName = $parsedClass->getName() . '::' . $reflectionConstant->getName();
+                yield $prefix . ' ' . $fullConstantName => [
+                    $parsedClass,
+                    $reflectionConstant,
+                ];
+            }
+        }
+    }
 
-        include_once $fileName;
+    /**
+     * @inheritDoc
+     */
+    static protected function getGettersToCheck(): array
+    {
+        $php83Getters = [];
+        if (PHP_VERSION_ID >= 80300) {
+            $php83Getters[] = 'hasType';
+        }
+        return [
+            'getDocComment', 'getModifiers', 'getName', 'getValue',
+            'isPrivate', 'isProtected', 'isPublic', 'isFinal', 'isEnumCase',
+            '__toString', ...$php83Getters
+        ];
     }
 }
