@@ -3,80 +3,44 @@ declare(strict_types=1);
 
 namespace Go\ParserReflection;
 
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
-class ReflectionFunctionTest extends TestCase
+class ReflectionFunctionTest extends AbstractTestCase
 {
-    const STUB_FILE55 = '/Stub/FileWithFunctions55.php';
-    const STUB_FILE70 = '/Stub/FileWithFunctions70.php';
+    protected static string $reflectionClassToTest = \ReflectionFunction::class;
+    protected const DEFAULT_STUB_FILENAME = '/Stub/FileWithFunctions55.php';
+    public const STUB_FILE70 = '/Stub/FileWithFunctions70.php';
 
     /**
-     * @var ReflectionFile
+     * Performs method-by-method comparison with original reflection
+     *
+     * @param ReflectionFunction  $parsedFunction Parsed function
+     * @param \ReflectionFunction $refFunction Original function
+     * @param string              $getterName Name of the reflection method to test
      */
-    protected $parsedRefFile;
+    #[DataProvider('reflectionGetterDataProvider')]
+    public function testReflectionGetterParity(
+        ReflectionFunction  $parsedFunction,
+        \ReflectionFunction $refFunction,
+        string              $getterName
+    ): void {
+        $functionName  = $refFunction->getName();
 
-    protected function setUp(): void
-    {
-        $fileName = stream_resolve_include_path(__DIR__ . self::STUB_FILE55);
-
-        $reflectionFile = new ReflectionFile($fileName);
-        $this->parsedRefFile = $reflectionFile;
-
-        include_once $fileName;
+        $expectedValue = $refFunction->$getterName();
+        $actualValue   = $parsedFunction->$getterName();
+        // I would like to completely stop maintaining the __toString method
+        if ($expectedValue !== $actualValue && $getterName === '__toString') {
+            $this->markTestSkipped("__toString for function {$functionName}() is not equal:\n{$expectedValue}{$actualValue}");
+        }
+        $this->assertSame(
+            $expectedValue,
+            $actualValue,
+            "$getterName() for function {$functionName}() should be equal"
+        );
     }
 
-    public function testGeneralInfoGetters()
-    {
-        $allNameGetters = [
-            'getStartLine', 'getEndLine', 'getDocComment', 'getExtension', 'getExtensionName',
-            'getName', 'getNamespaceName', 'getShortName', 'inNamespace', 'getStaticVariables',
-            'getNumberOfParameters', 'getNumberOfRequiredParameters', '__toString', 'isDisabled',
-            'returnsReference', 'getClosureScopeClass', 'getClosureThis'
-        ];
-
-        if (PHP_VERSION_ID >= 70000) {
-            $allNameGetters[] = 'hasReturnType';
-        }
-
-        foreach ($this->parsedRefFile->getFileNamespaces() as $fileNamespace) {
-            foreach ($fileNamespace->getFunctions() as $refFunction) {
-                $functionName        = $refFunction->getName();
-                $originalRefFunction = new \ReflectionFunction($functionName);
-                foreach ($allNameGetters as $getterName) {
-                    $expectedValue = $originalRefFunction->$getterName();
-                    $actualValue   = $refFunction->$getterName();
-                    $this->assertSame(
-                        $expectedValue,
-                        $actualValue,
-                        "{$getterName}() for function {$functionName} should be equal"
-                    );
-                }
-            }
-        }
-    }
-
-    public function testCoverAllMethods()
-    {
-        $allInternalMethods = get_class_methods(\ReflectionFunction::class);
-        $allMissedMethods   = [];
-
-        foreach ($allInternalMethods as $internalMethodName) {
-            if ('export' === $internalMethodName) {
-                continue;
-            }
-            $refMethod    = new \ReflectionMethod(ReflectionFunction::class, $internalMethodName);
-            $definerClass = $refMethod->getDeclaringClass()->getName();
-            if (strpos($definerClass, 'Go\\ParserReflection') !== 0) {
-                $allMissedMethods[] = $internalMethodName;
-            }
-        }
-
-        if ($allMissedMethods) {
-            $this->markTestIncomplete('Methods ' . join($allMissedMethods, ', ') . ' are not implemented');
-        }
-    }
-
-    public function testGetClosureMethod()
+    public function testGetClosureMethod(): void
     {
         $fileNamespace = $this->parsedRefFile->getFileNamespace('Go\ParserReflection\Stub');
         $refFunc       = $fileNamespace->getFunction('noGeneratorFunc');
@@ -84,64 +48,76 @@ class ReflectionFunctionTest extends TestCase
 
         $this->assertInstanceOf(\Closure::class, $closure);
         $retValue = $closure();
-        $this->assertEquals(100, $retValue);
+        $this->assertSame(100, $retValue);
     }
 
-    public function testInvokeMethod()
+    public function testInvokeMethod(): void
     {
         $fileNamespace = $this->parsedRefFile->getFileNamespace('Go\ParserReflection\Stub');
         $refFunc       = $fileNamespace->getFunction('funcWithReturnArgs');
         $retValue      = $refFunc->invoke(1, 2, 3);
-        $this->assertEquals([1, 2, 3], $retValue);
+        $this->assertSame([1, 2, 3], $retValue);
     }
 
-    public function testInvokeArgsMethod()
+    public function testInvokeArgsMethod(): void
     {
         $fileNamespace = $this->parsedRefFile->getFileNamespace('Go\ParserReflection\Stub');
         $refFunc       = $fileNamespace->getFunction('funcWithReturnArgs');
         $retValue      = $refFunc->invokeArgs([1, 2, 3]);
-        $this->assertEquals([1, 2, 3], $retValue);
+        $this->assertSame([1, 2, 3], $retValue);
     }
 
-    public function testGetReturnTypeMethod()
-    {
-        if (PHP_VERSION_ID < 70000) {
-            $this->markTestSkipped('Test available only for PHP7.0 and newer');
+    #[DataProvider('functionsDataProvider')]
+    public function testGetReturnTypeMethod(
+        ReflectionFunction $parsedRefFunction, \ReflectionFunction $originalRefFunction
+    ): void {
+        $functionName  = $parsedRefFunction->getName();
+        $hasReturnType = $originalRefFunction->hasReturnType();
+        $this->assertSame(
+            $originalRefFunction->hasReturnType(),
+            $parsedRefFunction->hasReturnType(),
+            "Presence of return type for function {$functionName} should be equal"
+        );
+        if ($hasReturnType) {
+            $parsedReturnType   = $parsedRefFunction->getReturnType();
+            $originalReturnType = $originalRefFunction->getReturnType();
+            $this->assertSame($originalReturnType->allowsNull(), $parsedReturnType->allowsNull());
+            $this->assertSame($originalReturnType->__toString(), $parsedReturnType->__toString());
+        } else {
+            $this->assertSame(
+                $originalRefFunction->getReturnType(),
+                $parsedRefFunction->getReturnType()
+            );
         }
+    }
 
-        $fileName = stream_resolve_include_path(__DIR__ . self::STUB_FILE70);
-
-        $reflectionFile = new ReflectionFile($fileName);
-        include $fileName;
-
-        foreach ($reflectionFile->getFileNamespaces() as $fileNamespace) {
-            foreach ($fileNamespace->getFunctions() as $refFunction) {
-                $functionName        = $refFunction->getName();
-                $originalRefFunction = new \ReflectionFunction($functionName);
-                $hasReturnType       = $refFunction->hasReturnType();
-                $this->assertSame(
-                    $originalRefFunction->hasReturnType(),
-                    $hasReturnType,
-                    "Presence of return type for function {$functionName} should be equal"
-                );
-                if ($hasReturnType) {
-                    $parsedReturnType   = $refFunction->getReturnType();
-                    $originalReturnType = $originalRefFunction->getReturnType();
-                    $this->assertSame($originalReturnType->allowsNull(), $parsedReturnType->allowsNull());
-                    $this->assertSame($originalReturnType->isBuiltin(), $parsedReturnType->isBuiltin());
-                    // TODO: To prevent deprecation error in tests
-                    if (PHP_VERSION_ID < 70400) {
-                        $this->assertSame($originalReturnType->__toString(), $parsedReturnType->__toString());
-                    } else {
-                        $this->assertSame($originalReturnType->getName(), $parsedReturnType->__toString());
-                    }
-                } else {
-                    $this->assertSame(
-                        $originalRefFunction->getReturnType(),
-                        $refFunction->getReturnType()
-                    );
-                }
+    /**
+     * Provides full test-case list in the form [ReflectionFunction, \ReflectionFunction, getter name to check]
+     */
+    public static function reflectionGetterDataProvider(): \Generator
+    {
+        $allNameGetters = self::getGettersToCheck();
+        foreach (self::functionsDataProvider() as $prefix => [$parsedFunction, $refFunction]) {
+            foreach ($allNameGetters as $getterName) {
+                yield $prefix . ', ' . $getterName => [
+                    $parsedFunction,
+                    $refFunction,
+                    $getterName
+                ];
             }
         }
+    }
+
+    /**
+     * @inheritDoc
+     */
+    static protected function getGettersToCheck(): array
+    {
+        return [
+            'getStartLine', 'getEndLine', 'getDocComment', 'getExtension', 'getExtensionName',
+            'getName', 'getNamespaceName', 'getShortName', 'inNamespace', 'getStaticVariables',
+            'getNumberOfParameters', 'getNumberOfRequiredParameters',
+            'returnsReference', 'getClosureScopeClass', 'getClosureThis', 'hasReturnType', '__toString'
+        ];
     }
 }
