@@ -3,204 +3,96 @@ declare(strict_types=1);
 
 namespace Go\ParserReflection;
 
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 use Go\ParserReflection\Stub\Foo;
 use Go\ParserReflection\Stub\SubFoo;
-use TestParametersForRootNsClass;
 
-class ReflectionParameterTest extends TestCase
+class ReflectionParameterTest extends AbstractTestCase
 {
+    protected const DEFAULT_STUB_FILENAME = '/Stub/FileWithParameters55.php';
+
+    protected static string $reflectionClassToTest = \ReflectionParameter::class;
+
     /**
-     * @var ReflectionFile
+     * Performs method-by-method comparison with original reflection
      */
-    protected $parsedRefFile;
+    #[DataProvider('reflectionGetterDataProvider')]
+    public function testReflectionGetterParity(
+        \ReflectionFunctionAbstract $parsedFunctionAbstract,
+        ReflectionParameter $parsedParameter,
+        \ReflectionParameter $originalRefParameter,
+        string $getterName
+    ): void {
+        $parameterName = $originalRefParameter->getName();
+        if ($parsedFunctionAbstract instanceof \ReflectionMethod) {
+            $functionName = [$parsedFunctionAbstract->class, $parsedFunctionAbstract->getName()];
+        } else {
+            $functionName = $parsedFunctionAbstract->getName();
+        }
 
-    protected function setUp(): void
-    {
-        $this->setUpFile(__DIR__ . '/Stub/FileWithParameters55.php');
+        $expectedValue = $originalRefParameter->$getterName();
+        $actualValue   = $parsedParameter->$getterName();
+        $displayableName = is_array($functionName) ? join ('->', $functionName) : $functionName;
+        // I would like to completely stop maintaining the __toString method
+        if ($expectedValue !== $actualValue && $getterName === '__toString') {
+            $this->markTestSkipped("__toString for parameter {$displayableName}(\${$parameterName}) is not equal:\n{$expectedValue}\n{$actualValue}");
+        }
+        $this->assertSame(
+            $expectedValue,
+            $actualValue,
+            "{$getterName}() for parameter {$displayableName}(\${$parameterName}) should be equal"
+        );
+    }
+
+    #[DataProvider('parametersDataProvider')]
+    public function testGetClassMethod(
+        \ReflectionFunctionAbstract $parsedFunction,
+        ReflectionParameter $parsedParameter,
+        \ReflectionParameter        $originalRefParameter
+    ): void {
+        $originalParamClass = $originalRefParameter->getClass();
+        $parsedParamClass   = $parsedParameter->getClass();
+
+        if (isset($originalParamClass)) {
+            $this->assertNotNull($parsedParamClass, "Original param class is: {$originalParamClass->name}");
+            $this->assertSame($originalParamClass->getName(), $parsedParamClass->getName());
+        } else {
+            $this->assertNull($parsedParamClass);
+        }
+    }
+
+    #[DataProvider('parametersDataProvider')]
+    public function testGetDeclaringClassMethod(
+        \ReflectionFunctionAbstract $parsedFunction,
+        ReflectionParameter $parsedParameter,
+        \ReflectionParameter        $originalRefParameter
+    ): void {
+        $originalDeclaringClass = $originalRefParameter->getDeclaringClass();
+        $parsedDeclaringClass   = $parsedParameter->getDeclaringClass();
+
+        if (isset($originalDeclaringClass)) {
+            $this->assertSame($originalDeclaringClass->getName(), $parsedDeclaringClass->getName());
+        } else {
+            $this->assertNull($parsedDeclaringClass);
+        }
+    }
+
+    #[DataProvider('parametersDataProvider')]
+    public function testDebugInfoMethod(
+        \ReflectionFunctionAbstract $parsedFunction,
+        ReflectionParameter $parsedParameter,
+        \ReflectionParameter        $originalRefParameter
+    ): void {
+        $expectedValue  = (array) $originalRefParameter;
+        $this->assertSame($expectedValue, $parsedParameter->__debugInfo());
     }
 
     /**
-     * @dataProvider fileProvider
-     */
-    public function testGeneralInfoGetters($fileName)
-    {
-        $this->setUpFile($fileName);
-        $allNameGetters = [
-            'isArray', 'isCallable', 'isOptional', 'isPassedByReference', 'isDefaultValueAvailable',
-            'getPosition', 'canBePassedByValue', 'allowsNull', 'getDefaultValue', 'getDefaultValueConstantName',
-            'isDefaultValueConstant', '__toString'
-        ];
-        $onlyWithDefaultValues = array_flip([
-            'getDefaultValue', 'getDefaultValueConstantName', 'isDefaultValueConstant'
-        ]);
-        if (PHP_VERSION_ID >= 50600) {
-            $allNameGetters[] = 'isVariadic';
-        }
-        if (PHP_VERSION_ID >= 70000) {
-            $allNameGetters[] = 'hasType';
-        }
-
-        foreach ($this->parsedRefFile->getFileNamespaces() as $fileNamespace) {
-            foreach ($fileNamespace->getFunctions() as $refFunction) {
-                $functionName = $refFunction->getName();
-                foreach ($refFunction->getParameters() as $refParameter) {
-                    $parameterName        = $refParameter->getName();
-                    $originalRefParameter = new \ReflectionParameter($functionName, $parameterName);
-                    foreach ($allNameGetters as $getterName) {
-
-                        // skip some methods if there is no default value
-                        $isDefaultValueAvailable = $originalRefParameter->isDefaultValueAvailable();
-                        if (isset($onlyWithDefaultValues[$getterName]) && !$isDefaultValueAvailable) {
-                            continue;
-                        }
-                        $expectedValue = $originalRefParameter->$getterName();
-                        $actualValue   = $refParameter->$getterName();
-                        $this->assertSame(
-                            $expectedValue,
-                            $actualValue,
-                            "{$getterName}() for parameter {$functionName}:{$parameterName} should be equal"
-                        );
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * Provides a list of files for analysis
-     *
-     * @return array
-     */
-    public function fileProvider()
-    {
-        $files = ['PHP5.5' => [__DIR__ . '/Stub/FileWithParameters55.php']];
-
-        if (PHP_VERSION_ID >= 50600) {
-            $files['PHP5.6'] = [__DIR__ . '/Stub/FileWithParameters56.php'];
-        }
-        if (PHP_VERSION_ID >= 70000) {
-            $files['PHP7.0'] = [__DIR__ . '/Stub/FileWithParameters70.php'];
-        }
-
-        return $files;
-    }
-
-    public function testGetClassMethod()
-    {
-        $parsedNamespace = $this->parsedRefFile->getFileNamespace('Go\ParserReflection\Stub');
-        $parsedFunction  = $parsedNamespace->getFunction('miscParameters');
-
-        $parameters = $parsedFunction->getParameters();
-        $this->assertSame(null, $parameters[0 /* array $arrayParam*/]->getClass());
-        $this->assertSame(null, $parameters[3 /* callable $callableParam */]->getClass());
-
-        $objectParam = $parameters[5 /* \stdClass $objectParam */]->getClass();
-        $this->assertInstanceOf(\ReflectionClass::class, $objectParam);
-        $this->assertSame(\stdClass::class, $objectParam->getName());
-
-        $typehintedParamWithNs = $parameters[7 /* ReflectionParameter $typehintedParamWithNs */]->getClass();
-        $this->assertInstanceOf(\ReflectionClass::class, $typehintedParamWithNs);
-        $this->assertSame(ReflectionParameter::class, $typehintedParamWithNs->getName());
-
-        $internalInterfaceParam = $parameters[12 /* \Traversable $traversable */]->getClass();
-        $this->assertInstanceOf(\ReflectionClass::class, $internalInterfaceParam);
-        $this->assertSame(\Traversable::class, $internalInterfaceParam->getName());
-    }
-
-    public function testGetClassMethodReturnsSelfAndParent()
-    {
-        $parsedNamespace = $this->parsedRefFile->getFileNamespace('Go\ParserReflection\Stub');
-        $parsedClass     = $parsedNamespace->getClass(SubFoo::class);
-        $parsedFunction  = $parsedClass->getMethod('anotherMethodParam');
-
-        $parameters = $parsedFunction->getParameters();
-        $selfParam = $parameters[0 /* self $selfParam */]->getClass();
-        $this->assertInstanceOf(\ReflectionClass::class, $selfParam);
-        $this->assertSame(SubFoo::class, $selfParam->getName());
-
-        $parentParam = $parameters[1 /* parent $parentParam */]->getClass();
-        $this->assertInstanceOf(\ReflectionClass::class, $parentParam);
-        $this->assertSame(Foo::class, $parentParam->getName());
-    }
-
-    public function testNonConstantsResolvedForGlobalNamespace()
-    {
-        $parsedNamespace = $this->parsedRefFile->getFileNamespace('');
-        $parsedClass     = $parsedNamespace->getClass(TestParametersForRootNsClass::class);
-        $parsedFunction  = $parsedClass->getMethod('foo');
-
-        $parameters = $parsedFunction->getParameters();
-        $this->assertSame(null, $parameters[0]->getDefaultValue());
-        $this->assertSame(false, $parameters[1]->getDefaultValue());
-        $this->assertSame(true, $parameters[2]->getDefaultValue());
-    }
-
-    public function testGetDeclaringClassMethodReturnsObject()
-    {
-        $parsedNamespace = $this->parsedRefFile->getFileNamespace('Go\ParserReflection\Stub');
-        $parsedClass     = $parsedNamespace->getClass(Foo::class);
-        $parsedFunction  = $parsedClass->getMethod('methodParam');
-
-        $parameters = $parsedFunction->getParameters();
-        $this->assertSame($parsedClass->getName(), $parameters[0]->getDeclaringClass()->getName());
-    }
-
-    public function testParamWithDefaultConstValue()
-    {
-        $parsedNamespace = $this->parsedRefFile->getFileNamespace('Go\ParserReflection\Stub');
-        $parsedClass     = $parsedNamespace->getClass(Foo::class);
-        $parsedFunction  = $parsedClass->getMethod('methodParamConst');
-
-        $parameters = $parsedFunction->getParameters();
-        $this->assertTrue($parameters[0]->isDefaultValueConstant());
-        $this->assertSame('self::CLASS_CONST', $parameters[0]->getDefaultValueConstantName());
-
-        $this->assertTrue($parameters[2]->isDefaultValueConstant());
-        $this->assertSame('Go\ParserReflection\Stub\TEST_PARAMETER', $parameters[2]->getDefaultValueConstantName());
-
-        $this->assertTrue($parameters[3]->isDefaultValueConstant());
-        $this->assertSame('Go\ParserReflection\Stub\SubFoo::ANOTHER_CLASS_CONST', $parameters[3]->getDefaultValueConstantName());
-    }
-
-    public function testParamBuiltInClassConst()
-    {
-        $parsedNamespace = $this->parsedRefFile->getFileNamespace('Go\ParserReflection\Stub');
-        $parsedClass     = $parsedNamespace->getClass(Foo::class);
-        $parsedFunction  = $parsedClass->getMethod('methodParamBuiltInClassConst');
-
-        $parameters = $parsedFunction->getParameters();
-        $this->assertTrue($parameters[0]->isDefaultValueConstant());
-        $this->assertSame('DateTime::ATOM', $parameters[0]->getDefaultValueConstantName());
-    }
-
-    public function testGetDeclaringClassMethodReturnsNull()
-    {
-        $parsedNamespace = $this->parsedRefFile->getFileNamespace('Go\ParserReflection\Stub');
-        $parsedFunction  = $parsedNamespace->getFunction('miscParameters');
-
-        $parameters = $parsedFunction->getParameters();
-        $this->assertNull($parameters[0]->getDeclaringClass());
-    }
-
-    public function testDebugInfoMethod()
-    {
-        $parsedNamespace = $this->parsedRefFile->getFileNamespace('Go\ParserReflection\Stub');
-        $parsedFunction  = $parsedNamespace->getFunction('miscParameters');
-
-        $parsedRefParameters  = $parsedFunction->getParameters();
-        $parsedRefParameter   = $parsedRefParameters[0];
-        $originalRefParameter = new \ReflectionParameter('Go\ParserReflection\Stub\miscParameters', 'arrayParam');
-        $expectedValue        = (array) $originalRefParameter;
-        $this->assertSame($expectedValue, $parsedRefParameter->__debugInfo());
-    }
-
-    /**
-     * @dataProvider listOfDefaultGetters
-     *
      * @param string $getterName Name of the getter to call
      */
-    public function testGetDefaultValueThrowsAnException($getterName)
+    #[DataProvider('listOfDefaultGetters')]
+    public function testGetDefaultValueThrowsAnException(string $getterName): void
     {
         $originalException = null;
         $parsedException   = null;
@@ -228,91 +120,110 @@ class ReflectionParameterTest extends TestCase
         $this->assertSame($originalException->getMessage(), $parsedException->getMessage());
     }
 
-    public function listOfDefaultGetters()
+    public static function listOfDefaultGetters(): \Iterator
     {
-        return [
-            ['getDefaultValue'],
-            ['getDefaultValueConstantName']
+        yield ['getDefaultValue'];
+        yield ['getDefaultValueConstantName'];
+    }
+
+    #[DataProvider('parametersDataProvider')]
+    public function testGetTypeMethod(
+        \ReflectionFunctionAbstract $parsedFunctionAbstract,
+        ReflectionParameter $parsedParameter,
+        \ReflectionParameter $originalRefParameter
+    ): void {
+        $functionName  = $parsedFunctionAbstract->getName();
+        $parameterName = $parsedParameter->getName();
+        $hasType       = $originalRefParameter->hasType();
+        $this->assertSame(
+            $hasType,
+            $parsedParameter->hasType(),
+            "Presence of type for parameter {$functionName}:{$parameterName} should be equal"
+        );
+        $message= "Parameter $functionName:$parameterName not equals to the original reflection";
+        if ($hasType) {
+            $parsedReturnType   = $parsedParameter->getType();
+            $originalReturnType = $originalRefParameter->getType();
+            $this->assertSame($originalReturnType->allowsNull(), $parsedReturnType->allowsNull(), $message);
+            $this->assertSame($originalReturnType->__toString(), $parsedReturnType->__toString(), $message);
+        } else {
+            $this->assertSame(
+                $originalRefParameter->getType(),
+                $parsedParameter->getType(),
+                $message
+            );
+        }
+    }
+
+    /**
+     * Provides full test-case list in the form [ReflectionClass, \ReflectionMethod, getter name to check]
+     */
+    public static function reflectionGetterDataProvider(): \Generator
+    {
+        static $onlyWithDefaultValues = [
+            'getDefaultValue', 'getDefaultValueConstantName', 'isDefaultValueConstant'
         ];
-    }
 
-    public function testCoverAllMethods()
-    {
-        $allInternalMethods = get_class_methods(\ReflectionParameter::class);
-        $allMissedMethods   = [];
-
-        foreach ($allInternalMethods as $internalMethodName) {
-            if ('export' === $internalMethodName) {
-                continue;
-            }
-            $refMethod    = new \ReflectionMethod(ReflectionParameter::class, $internalMethodName);
-            $definerClass = $refMethod->getDeclaringClass()->getName();
-            if (strpos($definerClass, 'Go\\ParserReflection') !== 0) {
-                $allMissedMethods[] = $internalMethodName;
-            }
-        }
-
-        if ($allMissedMethods) {
-            $this->markTestIncomplete('Methods ' . join($allMissedMethods, ', ') . ' are not implemented');
-        }
-    }
-
-    public function testGetTypeMethod()
-    {
-        if (PHP_VERSION_ID < 70000) {
-            $this->markTestSkipped('Test available only for PHP7.0 and newer');
-        }
-        $this->setUpFile(__DIR__ . '/Stub/FileWithParameters70.php');
-
-        foreach ($this->parsedRefFile->getFileNamespaces() as $fileNamespace) {
-            foreach ($fileNamespace->getFunctions() as $refFunction) {
-                $functionName = $refFunction->getName();
-                foreach ($refFunction->getParameters() as $refParameter) {
-                    $parameterName        = $refParameter->getName();
-                    $originalRefParameter = new \ReflectionParameter($functionName, $parameterName);
-                    $hasType              = $refParameter->hasType();
-                    $this->assertSame(
-                        $originalRefParameter->hasType(),
-                        $hasType,
-                        "Presence of type for parameter {$functionName}:{$parameterName} should be equal"
-                    );
-                    $message= "Parameter $functionName:$parameterName not equals to the original reflection";
-                    if ($hasType) {
-                        $parsedReturnType   = $refParameter->getType();
-                        $originalReturnType = $originalRefParameter->getType();
-                        $this->assertSame($originalReturnType->allowsNull(), $parsedReturnType->allowsNull(), $message);
-                        $this->assertSame($originalReturnType->isBuiltin(), $parsedReturnType->isBuiltin(), $message);
-                        // TODO: To prevent deprecation error in tests
-                        if (PHP_VERSION_ID < 70400) {
-                            $this->assertSame($originalReturnType->__toString(), $parsedReturnType->__toString(), $message);
-                        } else {
-                            $this->assertSame($originalReturnType->getName(), $parsedReturnType->__toString(), $message);
-                        }
-                    } else {
-                        $this->assertSame(
-                            $originalRefParameter->getType(),
-                            $refParameter->getType(),
-                            $message
-                        );
-                    }
+        $allNameGetters = self::getGettersToCheck();
+        foreach (self::parametersDataProvider() as $prefix => [$parsedFunction, $parsedParameter, $originalParameter]) {
+            foreach ($allNameGetters as $getterName) {
+                // We should ignore some methods if there isn't default value
+                $isDefaultValueAvailable = $originalParameter->isDefaultValueAvailable();
+                if (!$isDefaultValueAvailable && in_array($getterName, $onlyWithDefaultValues)) {
+                    continue;
                 }
+                yield $prefix . ', ' . $getterName => [
+                    $parsedFunction,
+                    $parsedParameter,
+                    $originalParameter,
+                    $getterName
+                ];
             }
         }
     }
 
     /**
-     * Setups file for parsing
-     *
-     * @param string $fileName File name to use
+     * Provides generator list in the form [ReflectionFunctionAbstract, ReflectionParameter, \ReflectionParameter to check]
      */
-    private function setUpFile($fileName)
+    public static function parametersDataProvider(): \Generator
     {
-        $fileName = stream_resolve_include_path($fileName);
-        $fileNode = ReflectionEngine::parseFile($fileName);
+        foreach (self::methodsDataProvider() as $prefix => [$parsedClass, $parsedClassMethod]) {
+            if (get_class($parsedClassMethod) === \ReflectionMethod::class) {
+                // We don't want test again parent parameters from parent methods, which already loaded
+                continue;
+            }
+            foreach ($parsedClassMethod->getParameters() as $parsedMethodParameter) {
+                $paramName = $parsedMethodParameter->getName();
+                $refParameter = new \ReflectionParameter([$parsedClass->getName(), $parsedClassMethod->getName()], $paramName);
+                yield $prefix . ' ' . '($' . $paramName . ')' => [
+                    $parsedClassMethod,
+                    $parsedMethodParameter,
+                    $refParameter
+                ];
+            }
+        }
+        foreach (self::functionsDataProvider() as $prefix => [$parsedFunction, $refFunction]) {
+            foreach ($parsedFunction->getParameters() as $parsedFunctionParameter) {
+                $paramName = $parsedFunctionParameter->getName();
+                $refParameter = new \ReflectionParameter($parsedFunction->getName(), $paramName);
+                yield $prefix . ' ' . '($' . $paramName . ')' => [
+                    $parsedFunction,
+                    $parsedFunctionParameter,
+                    $refParameter
+                ];
+            }
+        }
+    }
 
-        $reflectionFile = new ReflectionFile($fileName, $fileNode);
-        $this->parsedRefFile = $reflectionFile;
-
-        include_once $fileName;
+    /**
+     * @inheritDoc
+     */
+    static protected function getGettersToCheck(): array
+    {
+        return [
+            'isOptional', 'isPassedByReference', 'isDefaultValueAvailable',
+            'getPosition', 'canBePassedByValue', 'allowsNull', 'getDefaultValue', 'getDefaultValueConstantName',
+            'isDefaultValueConstant', 'isVariadic', 'isPromoted', 'hasType', '__toString'
+        ];
     }
 }
