@@ -11,13 +11,7 @@ declare(strict_types=1);
 
 namespace Go\ParserReflection\Stub;
 
-function miscScalarParameters(
-    int $acceptsInteger,
-    string $acceptsString,
-    bool $acceptsBool,
-    int $acceptsVariadicInteger,
-    array $acceptsArray = ['some' => ['value' => 'for', 'array' => 'to dump']],
-    float $acceptsFloat = \INF,
-    string ...$acceptsDefaultString
-) {
+#[SomeAttribute]
+function function_with_attribute()
+{
 }
