@@ -16,5 +16,5 @@ namespace Go\ParserReflection\Stub {
     {
     }
 
-    function builtInArgs(int $a, float $b, bool $c, string $d, object $e) : integer {}
+    function builtInArgs(int $a, float $b, bool $c, string $d, object $e) : int {}
 }
