@@ -11,13 +11,9 @@ declare(strict_types=1);
 
 namespace Go\ParserReflection\Stub;
 
-function miscScalarParameters(
-    int $acceptsInteger,
-    string $acceptsString,
-    bool $acceptsBool,
-    int $acceptsVariadicInteger,
-    array $acceptsArray = ['some' => ['value' => 'for', 'array' => 'to dump']],
-    float $acceptsFloat = \INF,
-    string ...$acceptsDefaultString
+function authenticate(
+    string $username,
+    #[\SensitiveParameter]
+    string $password
 ) {
 }
